feat(gallery): display image from request
