+++ package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "body-parser": "^1.18.3",
+    "cors": "^2.8.5",
     "express": "^4.16.4",
     "foreman": "^3.0.1",
     "jest": "^24.5.0",
